requirements.txt
asttokens==2.4.1
blinker==1.7.0
click==8.1.7
decorator==5.1.1
executing==2.0.1
Flask==3.0.0
greenlet==3.0.1
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
prompt-toolkit==3.0.41
psycopg==3.1.12
psycopg-binary==3.1.12
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.8.0
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.23
stack-data==0.6.3
traitlets==5.13.0
typing_extensions==4.8.0
wcwidth==0.2.10
Werkzeug==3.0.1
